refactor reducer and constants to separate file
